+++ CHANGELOG.md
@@ -18,6 +18,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.337.0 - 2023-10-30
+
+### Changed
+
+- 施設を最適化
+- 座席の$projectionを定義
+- aws-sdk -> @aws-sdk/client-cognito-identity-provider
+- update typescript@5.x.x
+
+### Removed
+
+- 注文の非推奨決済方法区分コードを廃止
+
 ## v4.336.0 - 2023-10-26
 
 ### Changed